pyproject.toml
[project]
name = "websearchagent"
version = "0.1.0"
description = ""
authors = [
    {name = "LaljiSankhat03",email = "laljis@webelight.co.in"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "dotenv (>=0.9.9,<0.10.0)",
    "google-search-results (>=2.4.2,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "langchain-groq (>=1.1.1,<2.0.0)",
    "langgraph[postgres] (>=1.0.6,<2.0.0)",
    "langchain-text-splitters (>=1.1.0,<2.0.0)",
    "langchain (>=1.2.6,<2.0.0)",
    "tavily-python (>=0.7.19,<0.8.0)",
    "langgraph-checkpoint-postgres (>=3.0.3,<4.0.0)"
]

[tool.poetry]
packages = [{include = "websearchagent", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
